% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{New Package}
\usage{
hello(x)
}
\arguments{
\item{x}{name of the person}
}
\value{
the output from \code{\link{print}}
}
\description{
Privde a functionality
}
\examples{
hello("John")
\dontrun{
hello("peter")
}
}
